+++ classes/capitalize.rb
@@ -1,0 +1,7 @@
+require_relative './decorator'
+
+class CapitalizeDecoration < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
