+++ classes/classroom.rb
@@ -2,7 +2,7 @@ require_relative './student'
 require_relative '../modules/hooky'
 
 class Classroom
-  include Play_hookey
+  include PlayHookey
 
   attr_accessor :label
   attr_reader :students
