+++ classes/trimme.rb
@@ -1,0 +1,7 @@
+require_relative './decorator'
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.slice(0, 9)
+  end
+end
